Refactor LiqPay API client for improved readability and Laravel 9 & 10 support
The LiqPay API PHP client constructor and other methods have been refactored for improved readability and easier understanding. This includes changes in the way constructor parameters are declared, and the implementation of other class methods.

Furthermore, explicit support for Laravel versions 9 and 10 has been added to ensure compatibility with the newer versions of the Laravel framework.